refactoring seglog file structure + reusing frame type
